fix: port the function to utils
