+++ src/utils.ts
@@ -5,6 +5,9 @@ import type {
 import type {
   HubInfo
 } from '../src/common/relay.types';
+import type HttpClient from './api/common/HttpClient';
+import {getDefaultHttpClient} from './api/common/HttpClient';
+import config from 'config';
 
 export const isDeployTransaction = (
   req: RelayTransactionRequest | DeployTransactionRequest
@@ -98,4 +101,34 @@ export interface EnvelopingTransactionDetails {
 
   retries?: number;
   initialBackoff?: number;
+}
+
+export const ENVELOPING_CONFIG_FIELD_NAME = 'EnvelopingConfig';
+export const PREFERRED_RELAYS_FIELD_NAME = 'preferredRelays';
+
+export async function selectNextRelay(
+  httpClient: HttpClient = getDefaultHttpClient()
+): Promise<RelayInfo | undefined> {
+  const preferredRelays: Array<RelayManagerData> = config.get(
+    `${ENVELOPING_CONFIG_FIELD_NAME}.${PREFERRED_RELAYS_FIELD_NAME}`
+  );
+
+  for (const relayInfo of preferredRelays) {
+    let hubInfo;
+
+    try{
+        hubInfo = await httpClient.getChainInfo(relayInfo.url);
+    }catch(error){
+        continue;
+    }
+
+    if (hubInfo.ready) {
+      return {
+        hubInfo,
+        relayInfo,
+      };
+    }
+  }
+
+  return undefined;
 }
